move configurePlugin function to separate module. pass initial config to separate module.
